+++ Shared/src/main/java/hu/szakdolgozat/targyak/Bot.java
@@ -4,7 +4,7 @@ import hu.szakdolgozat.Pozicio;
 import hu.szakdolgozat.TerkepKod;
 
 public class Bot extends Targy{
-    public static final int ID = TerkepKod.BOT;
+    public static final long ID = TerkepKod.BOT;
 
     public Bot(Pozicio pozicio) {
         super(ID, pozicio);
