+++ Shared/src/main/java/hu/szakdolgozat/targyak/Level.java
@@ -4,7 +4,7 @@ import hu.szakdolgozat.Pozicio;
 import hu.szakdolgozat.TerkepKod;
 
 public class Level extends Targy {
-    public static final int ID = TerkepKod.LEVEL;
+    public static final long ID = TerkepKod.LEVEL;
 
     public Level(Pozicio pozicio) {
         super(ID, pozicio);
