+++ Shared/src/main/java/hu/szakdolgozat/targyak/Uveg.java
@@ -4,7 +4,7 @@ import hu.szakdolgozat.Pozicio;
 import hu.szakdolgozat.TerkepKod;
 
 public class Uveg extends Targy {
-    public static final int ID = TerkepKod.UVEG;
+    public static final long ID = TerkepKod.UVEG;
 
     public Uveg(Pozicio pozicio) {
         super(ID, pozicio);
